pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Victor Ikuomola <vheecthurikuomola@gmail.com>"]
description = ""
name = "llm-benchmark-sim"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
APScheduler = "^3.10.4"
SQLAlchemy = "^2.0.36"
alembic = "^1.13.3"
fastapi = "^0.115.3"
httpx = "^0.27.2"
numpy = "^2.1.2"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
python = "^3.12"
python-dotenv = "^1.0.1"
redis = "^5.1.1"
uvicorn = "^0.32.0"

[tool.poetry.dev-dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
toml-sort = "^0.23.1"
